# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BasicResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object"
			},
			"GoodbyeResponse": {
				"properties": {
					"date": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"message"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "code-verifier-back",
		"version": "1.0.0",
		"description": "backend node + express + ts + swagger",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@matiSal"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/hello": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise of Basicresponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint to retreive a message \"hello {name}\" in JSON",
				"tags": [
					"HelloController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Name of user to be greeted",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/goodbye": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise of Goodbyeresponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GoodbyeResponse"
								}
							}
						}
					}
				},
				"description": "Endpoint to retreive a message \"hello {name}\" in JSON",
				"tags": [
					"GoodbyeController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Name of user to be greeted",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "all users o user found by id",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retreive the users in the collection 'Users' of DB",
				"tags": [
					"userController"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of user to retreive (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "message info if deletion was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to delete the users in the collection 'Users' of DB",
				"tags": [
					"userController"
				],
				"security": [],
				"parameters": [
					{
						"description": "id of user to delete (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "message info if CREATED was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to create the users in the collection 'Users' of DB",
				"tags": [
					"userController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "user",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "message info if updated was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to update the users in the collection 'Users' of DB",
				"tags": [
					"userController"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of user to UPDATE",
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "user",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}